sync/requirements.txt
black==23.1.0
click==8.1.3
gitdb==4.0.10
GitPython==3.1.30
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.0
platformdirs==3.0.0
smmap==5.0.0
